Improvements: 1. better LoadedPageFormatter 2. better updateFetchSchedule

+++ pulsar-skeleton/src/main/kotlin/ai/platon/pulsar/crawl/fetch/TaskScheduler.kt
@@ -299,7 +299,7 @@ class TaskScheduler(
         enumCounters.inc(CommonCounter.rLinks, page.impreciseLinkCount)
 
         totalPages.incrementAndGet()
-        totalBytes.addAndGet(page.contentLength.toLong())
+        totalBytes.addAndGet(page.contentLength)
 
         if (page.isSeed) {
             enumCounters.inc(Counter.rSeeds)
